pyproject.toml
[project]
name = "ai-security-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "pygithub>=2.9.0",
    "python-dotenv>=1.2.2",
    "tabulate>=0.10.0",
    "uvicorn>=0.43.0",
]
